requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.13.2
structlog==25.4.0
python-dotenv==1.1.0
httpx==0.28.1
langfuse==3.2.1
pytest==8.3.5
streamlit
openai